+++ spec/system/run_test_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe "Running built-in tests via CLI" do
   subject(:test_run) { Dicey::CLI.call(%w[--test]) }
 
   it "exits with true and outputs test results" do
-    expect { test_run }.to output(/\A#{<<~TEXT}.+^\("s","a",4\)\+\("s","a",4\):/m).to_stdout
+    expect { test_run }.to output(/\A#{<<~TEXT}.+^\(s,a,4\)\+\(s,a,4\):/m).to_stdout
       D1:
         Dicey::DistributionCalculators::Trivial: ✔
         Dicey::DistributionCalculators::Binomial: ☂
@@ -27,7 +27,7 @@ RSpec.describe "Running built-in tests via CLI" do
     end
 
     it "completes successfully, skipping non-numeric dice tests" do
-      expect { test_run }.not_to output(a_string_including(%{("s","a",4)+("s","a",4):})).to_stdout
+      expect { test_run }.not_to output(a_string_including(%{(s,a,4)+(s,a,4):})).to_stdout
       expect(test_run).to be true
     end
   end